move Logger to utils
